+++ cosmic-news/src/main/java/es/codeurjc/cosmic_news/restcontroller/NewsRestController.java
@@ -25,23 +25,31 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import es.codeurjc.cosmic_news.DTO.NewsDTO;
 import es.codeurjc.cosmic_news.model.News;
+import es.codeurjc.cosmic_news.model.User;
 import es.codeurjc.cosmic_news.service.NewsService;
+import es.codeurjc.cosmic_news.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/api/news")
 public class NewsRestController {
     @Autowired
     NewsService newsService;
+
+	@Autowired
+    UserService userService;
     
      @Operation(summary = "Get paged news.")
     @ApiResponses(value = {
@@ -50,11 +58,18 @@ public class NewsRestController {
         @ApiResponse(responseCode = "204", description = "News not found, probably high page number supplied", content = @Content)
     })
     @GetMapping
-    public ResponseEntity<List<NewsDTO>> getNewsPage(
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "5") int size) {
+    public ResponseEntity<List<NewsDTO>> getNewsPage(@RequestParam(defaultValue = "0") int page, @RequestParam(defaultValue = "5") int size, @RequestParam(defaultValue = "date") String filter) {
 
-        Page<News> newsPage = newsService.findAll(PageRequest.of(page, size));
+		Page<News> newsPage = null;
+		if (filter.equals("likes")){
+            newsPage = newsService.findAllByLikes(PageRequest.of(page, size));
+        }else if(filter.equals("date")){
+            newsPage = newsService.findAllByDate(PageRequest.of(page, size));
+        }else if(filter.equals("time")){
+            newsPage = newsService.findAllByTime(PageRequest.of(page, size));
+        }else{
+            newsPage = newsService.findAll(PageRequest.of(page, size));
+        }
         List<NewsDTO> newsList = new ArrayList<>();
 
         for (News news : newsPage) {
@@ -67,7 +82,7 @@ public class NewsRestController {
         return ResponseEntity.status(HttpStatus.OK).body(newsList);
     }
 
-        @Operation(summary = "Get a news by its id")
+    @Operation(summary = "Get a news by its id")
 	@ApiResponses(value = {
 	 @ApiResponse(
 	 responseCode = "200",
@@ -110,7 +125,8 @@ public class NewsRestController {
 	})
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteNews(@PathVariable long id, Principal principal){
-        if(principal !=null){
+		HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        if(principal !=null && request.isUserInRole("ADMIN")){
 			News news = newsService.findNewsById(id);
 			if (news != null){
 				newsService.deleteNews(news.getId());
@@ -133,14 +149,22 @@ public class NewsRestController {
 	 responseCode = "400",
 	 description = "Data entered incorrectly.",
 	 content = @Content
-	 )
+	 ),
+	 @ApiResponse(
+		responseCode = "401",
+		description = "You are not authorized",
+		content = @Content
+		)
 	})
     @PostMapping()
 	@ResponseStatus(HttpStatus.CREATED)
-	public ResponseEntity<News> createNews(@RequestBody NewsDTO newsDTO) {
-        News news = newsDTO.toNews();
-		newsService.saveNews(news);
-		return new ResponseEntity<>(news, HttpStatus.OK);
+	public ResponseEntity<News> createNews(@RequestBody NewsDTO newsDTO, Principal principal) {
+		HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        if(principal !=null && request.isUserInRole("ADMIN")){
+			News news = newsDTO.toNews();
+			newsService.saveNews(news);
+			return new ResponseEntity<>(news, HttpStatus.OK);
+		}else return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 	}
 
     @Operation(summary = "Update a news fields by ID. You need to be an administrator.")
@@ -171,13 +195,16 @@ public class NewsRestController {
 	})
     @PutMapping("/{id}")
 	public ResponseEntity<News> updateNews(@PathVariable long id, @RequestBody NewsDTO newsDTO, Principal principal) throws SQLException {
-        News news = newsService.findNewsById(id);
+		HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        if(principal !=null && request.isUserInRole("ADMIN")){
+			News news = newsService.findNewsById(id);
 			if (news != null) {
-                updateNews(news, newsDTO);
+				updateNews(news, newsDTO);
 				return new ResponseEntity<>(news, HttpStatus.OK);
 			} else	{
 				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 			}
+		}else return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 	}
 
     @Operation(summary = "Post a new news of a picture by id")
@@ -201,7 +228,8 @@ public class NewsRestController {
     @PostMapping("/{id}/image")
 	public ResponseEntity<Object> uploadPhoto(@PathVariable long id, @RequestParam MultipartFile imageFile, Principal principal)
 			throws IOException {
-		if(principal !=null){
+		HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        if(principal !=null && request.isUserInRole("ADMIN")){
             News news = newsService.findNewsById(id);
 			if (news != null) {
 				URI location = ServletUriComponentsBuilder.fromCurrentRequest().build().toUri();
@@ -271,7 +299,8 @@ public class NewsRestController {
 	})
 	@DeleteMapping("/{id}/image")
 	public ResponseEntity<Object> deletePhoto(@PathVariable long id, Principal principal) throws IOException {
-		if(principal !=null){
+		HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
+        if(principal !=null && request.isUserInRole("ADMIN")){
             News news = newsService.findNewsById(id);
 			if (news != null) {
 				news.setPhoto(null);
@@ -284,6 +313,40 @@ public class NewsRestController {
 		}else return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 	}
 
+	@Operation(summary = "Give a like to the news by id. If you already gave like to this news, this will be a unlike")
+	@ApiResponses(value = {
+	 @ApiResponse(
+	 responseCode = "200",
+	 description = "Like/Unlike given",
+	 content = @Content
+	 ),
+	 @ApiResponse(
+	 responseCode = "404",
+	 description = "Data entered incorrectly, probably invalid id supplied",
+	 content = @Content
+	 ),	 
+	 @ApiResponse(
+	 responseCode = "401",
+	 description = "You are not authorized",
+	 content = @Content
+	 )
+	})
+	@PostMapping("/{id}/like")
+	public ResponseEntity<News> likeNews(@PathVariable long id, Principal principal) {
+		if (principal != null) {
+            User user = userService.findUserByMail(principal.getName());
+            News news = newsService.findNewsById(id);
+            if (user != null && news != null){
+                newsService.like(news,user);
+				return new ResponseEntity<>(news, HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        }else{
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+	}
+	
     public void updateNews(News news, NewsDTO newsDTO){
 		if (newsDTO.getTitle()!=null) news.setTitle(newsDTO.getTitle());
         if (newsDTO.getAuthor()!=null) news.setAuthor(newsDTO.getAuthor());
